move store process to separate file
